+++ packages/constraint-engine/src/json-schema/referenceBuilders/buildDerivedTypedJsonReference.ts
@@ -1,4 +1,4 @@
-import { DerivedReferenceBuilder } from '../../types/builder';
+import { DerivedReferenceBuilder } from '../../types/builders/derivedReferenceBuilder';
 import { UnknownTargetPath } from '../../types/targetPath';
 import { TargetReference } from '../../types/targetReference';
 import { JsonMetaTargetTypeId } from '../types/constants';
