+++ packages/constraint-engine/src/types/builders/derivedReferenceSetBuilder.ts
@@ -1,0 +1,12 @@
+import { UnknownTargetPath } from '../targetPath';
+import { TargetReference } from '../targetReference';
+import { UnknownTypedTarget } from '../typedTarget';
+
+export type DerivedReferenceSetBuilder<
+  TInputTypedTarget extends UnknownTypedTarget,
+  TInputTargetPath extends UnknownTargetPath,
+  TOutputTypedTarget extends UnknownTypedTarget,
+  TOutputTargetPath extends UnknownTargetPath,
+> = (
+  inputReference: TargetReference<TInputTypedTarget, TInputTargetPath>,
+) => TargetReference<TOutputTypedTarget, TOutputTargetPath>[];
