+++ packages/constraint-engine/src/types/builders/instanceBuilder.ts
@@ -1,0 +1,6 @@
+import { UnknownTargetInstance } from '../targetInstance';
+
+export type InstanceBuilder<
+  TInputData extends UnknownTargetInstance,
+  TOutputTargetInstance extends UnknownTargetInstance,
+> = (inputData: TInputData) => TOutputTargetInstance;
