+++ packages/constraint-engine/src/types/builders/referenceSetBuilder.ts
@@ -1,0 +1,12 @@
+import { UnknownTargetInstance } from '../targetInstance';
+import { UnknownTargetPath } from '../targetPath';
+import { TargetReference } from '../targetReference';
+import { UnknownTypedTarget } from '../typedTarget';
+
+export type ReferenceSetBuilder<
+  TInputData extends UnknownTargetInstance,
+  TOutputTypedTarget extends UnknownTypedTarget,
+  TOutputTargetPath extends UnknownTargetPath,
+> = (
+  inputData: TInputData,
+) => TargetReference<TOutputTypedTarget, TOutputTargetPath>[];
